Làm bài đi dm

+++ DeliveryMangementSystem/GUI/Aministrator.cs
@@ -4,6 +4,8 @@ using System.Windows.Forms;
 using System.Data;
 using System.Linq;
 using DeliveryMangementSystem.Forms.Reusable_Control;
+using Microsoft.EntityFrameworkCore;
+using DeliveryMangementSystem.Models;
 namespace DeliveryMangementSystem.Forms
 {
     public partial class frmAdmin : Form
@@ -14,12 +16,19 @@ namespace DeliveryMangementSystem.Forms
         //Methods
         private void LoadDataGridViewAccounts()
         {
-            //Load ds tài khoản (trừ admin) vào datagridview
+            //Load ds tài khoản vào datagridview
             using (var db = new myDbContext())
             {
-                var accounts = db.Accounts.Where(o => o.Role != Models.UserRole.Admin).ToList();
+                var accounts = db.Accounts.Include(a => a.Shipper).Where(s => s.Role != UserRole.Admin).Select(a => new
+                {
+                    a.Account_ID,
+                    Username = a.Username,
+                    Name = a.Shipper != null ? a.Shipper.Name : "",
+                    a.Role,
+                    a.Status
+                }).ToList();
                 dgvAccounts.DataSource = accounts;
-                
+         
             }
             FormatDataGridView();
         }
@@ -32,15 +41,15 @@ namespace DeliveryMangementSystem.Forms
             dgvAccounts.Columns.Add(new DataGridViewTextBoxColumn
             {
                 Name = "ID",
-                DataPropertyName = "Id",
+                DataPropertyName = "Account_ID",
                 HeaderText = "Mã tài khoản",
                 AutoSizeMode = DataGridViewAutoSizeColumnMode.Fill
             });
             dgvAccounts.Columns.Add(new DataGridViewTextBoxColumn
             {
-                Name = "Username",
-                DataPropertyName = "Username",
-                HeaderText = "Tên đăng nhập",
+                Name = "Name",
+                DataPropertyName = "Name",
+                HeaderText = "Chủ tài khoản",
                 AutoSizeMode = DataGridViewAutoSizeColumnMode.Fill
             });
             dgvAccounts.Columns.Add(new DataGridViewTextBoxColumn
@@ -48,6 +57,13 @@ namespace DeliveryMangementSystem.Forms
                 Name = "Role",
                 DataPropertyName = "Role",
                 HeaderText = "Vai trò",
+                AutoSizeMode = DataGridViewAutoSizeColumnMode.Fill
+            });
+            dgvAccounts.Columns.Add(new DataGridViewTextBoxColumn
+            {
+                Name = "Status",
+                DataPropertyName = "Status",
+                HeaderText = "Trạng thái tài khoản",
                 AutoSizeMode = DataGridViewAutoSizeColumnMode.Fill
             });
         }
@@ -79,14 +95,14 @@ namespace DeliveryMangementSystem.Forms
         {
             tabControlAdmin.SelectedTab = tpBranch;
         }
-
         private void btnChange_Click(object sender, EventArgs e)
         {
+            tabControlAdmin.SelectedTab = tpChangePassword;
             UC_ProfileChanger uC_ProfileChanger = new UC_ProfileChanger(Id, "Admin")
             {
                 Dock = DockStyle.Fill
             };
-            panel2.Controls.Add(uC_ProfileChanger);
+            tpChangePassword.Controls.Add(uC_ProfileChanger);
             uC_ProfileChanger.BringToFront();
         }
 
@@ -106,13 +122,13 @@ namespace DeliveryMangementSystem.Forms
 
                 using (var db = new myDbContext())
                 {
-                    var account = db.Accounts.FirstOrDefault(a => a.Id == accountId);
-                    if (account != null && account.ShipperId != null)
+                    var account = db.Accounts.FirstOrDefault(a => a.Account_ID == accountId);
+                    if (account != null && account.S_ID != null)
                     {
-                        var shipper = db.Shippers.Find(account.ShipperId);
+                        var shipper = db.Shippers.Find(account.S_ID);
                         if (shipper != null)
                         {
-                            var frm = new frmAccountDetail(shipper.Id);
+                            var frm = new frmAccountDetail(shipper.Shipper_ID);
                             frm.ShowDialog();
                         }
                     }
@@ -136,9 +152,16 @@ namespace DeliveryMangementSystem.Forms
                         DialogResult result = MessageBox.Show($"Bạn có chắc chắn muốn xóa tài khoản {accountId}?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
                         if (result == DialogResult.Yes)
                         {
-                            db.Accounts.Remove(account);
-                            db.SaveChanges();
-                            LoadDataGridViewAccounts();
+                            if (account.Status == UserStatus.Active)
+                            {
+                                MessageBox.Show("Tài khoản đang hoạt động, không thể xóa!");
+                            }
+                            else
+                            {
+                                db.Accounts.Remove(account);
+                                db.SaveChanges();
+                                LoadDataGridViewAccounts();
+                            }
                         }
                     }
                 }
